Create README.md
Demonstrates use of pure dependency injection. Code is nicely organized by following the classic MVC design pattern whereby the Activity/Fragment is controller and through it flow the use-cases. 
Next branch will show how to easily refactor to Dagger2 by easily replacing the CompositionRoot and ControllerCompositionRoot classes with Component classes, thus demonstrating that DI is an architectural best practice and Dagger2 is a framework which helps to easily manage the same.